different scale axis ratio (AR)  comparison
Because we set the monitor frequency interval to 0.5GHz during the simulation, some data distortion is caused.  However, in the actual measurement, our frequency interval is 0.01GHz, so there is a deviation between the actual measurement data and the simulation results, so we analyzed the simulated results of the 0.01GHz interval monitor. Finally, the comparison figure of simulation AR between 0.5GHz interval and 0.01GHz scales is uploaded